Move component files into the components directory
