+++ widgets/htdocs/alphafold/components/controlPanel.js
