+++ widgets/htdocs/alphafold/components/defaultColorsPanel.js
