+++ widgets/htdocs/alphafold/components/exonsControlPanel.js
