+++ widgets/htdocs/alphafold/components/proteinFeaturesControlPanel.js
