+++ widgets/htdocs/alphafold/components/variantsControlPanel.js
